EVM-IT: Move fevm.rs into fevm_features
