+++ testing/integration/Cargo.toml
@@ -69,8 +69,8 @@ tokio = { version = "1.24", features = ["rt-multi-thread", "macros"] }
 
 
 [[test]]
-name = "fevm"   # The EVM tests are using Cucumber
-harness = false # allows Cucumber to print output instead of libtest
+name = "fevm_features" # The EVM tests are using Cucumber
+harness = false        # allows Cucumber to print output instead of libtest
 
 [features]
 default = ["smt"]
